elliptic-curve: add `ScalarBytes` type
Adds a type for representing `FieldBytes` which is always in range of
the curve's `Order`.

Does not require the `arithmetic` feature or a `ProjectiveArithmetic`
implementation, however when those are enabled it supports infallible
conversions to/from `Scalar<C>`.
